fix profiles and posts add messages if user / post don't exists

+++ src/app/profile/page.tsx
@@ -11,19 +11,20 @@ import { getPosts } from '@/services/Prisma/post/getPosts'
 import { parseUser } from '@/functions/parseUser'
 
 const Welcome = async (): Promise<ReactElement> => {
-	const user = await parseUser()
-	console.log(user)
+	const user = await parseUser(true)
 
-	const posts = await getPosts({ authorId: [exists(user?.id)] })
+	if (!user.data) return <Container>{user.message}</Container>
+
+	const posts = await getPosts({ authorId: [exists(user?.data.id)] })
 
 	return (
 		<Container>
-			<UserProfile postsCount={exists<number>(posts.data?.length)} selfProfile user={user} />
+			<UserProfile postsCount={exists<number>(posts.data?.length)} selfProfile user={user.data} />
 			<Box direction="row" alignItems="start" gap={8} className={styles.box}>
 				<Button appearance="primary" icon="add_circle" href="/post">Новый пост</Button>
 				<Button appearance="secondary" icon="delete" href="/profile/deleted">Удаленные</Button>
 			</Box>
-			<Posts controls posts={exists(posts.data)} />
+			<Posts controls posts={posts.data ? posts.data : []} />
 		</Container>
 	)
 }
